Created an 'admin' folder, moved AdminDashboard.js into it
